Finished work on self driving AI
The Self-driving AI is done. And while i wished i could have trained it to be faster, i'm coing to continue work on another AI project. My sorting AI will be tasked with sorting balls of different colors. The task should be fairly easy for both me and the AI.