SortingAI_0.3
Extensive changes to how balls are spawned and how the AI observe. The AI is no longer able to observe the last two balls spawned, but are instead observing the color of balls going through the entrysensor. External decision requester have been added for this test.

The changes have been made, since i wanted to be able to spawn balls faster