SortingAI_0.3 Done training
As mentioned before. The AI had seen substantial changes in this version. The AI was only able to observe it's own arm position and which ball last passed thorugh the entry sensor at the top. The results were, that even at a speed of one ball pr 0.2 sec, the AI was able to sort them, even tho it had trained with one ball ever 1 second