Backup before rolling back
I have a new strategy. As of now, i want to sort the balls in four different colors. This means 3 arms to sort the incoming balls. Each arm have two options, tilt arm right or tilt arm left.

As of now, i'm unable to teach the AI to sort 4 different colors of balls, but i am perfectly capable of training it to sort two! So, what i'm going to do, is let every arm act on its own, and not as a collective AI. Meaning, each AI only have to think about sorting the balls correct that it meets. So, i'm, gonna train an AI for each arm in the game. Meaning i'll have two train 3 AI's (or 7 if i choose to sort 8 colors instead of 4)